Added MO_matrix_quaternion for complex spinor
In order to evaluate the spinor wavefunction, every basis function now
needs 2 numbers, the complex spin up and down coefficients. Due to the
way the orb file is read, it was convienient to incorporate a new
structure, a quaternion which can be represented as an ordered pair
of complex numbers. It is read in from an orbfile that has
coefficients that look like (a,b,c,d) to form the the spin up coeff
a+ib and the spin down c+id. Then the value of the contribution to the
molecular orbital coming from a particular basis function is
((a+ib)*exp(i*s) + (c+id)*exp(-i*s))*basis_func_val