Changes to Sample point
In sample point, all position arrays can either be [r,r2,x,y,z] or
[r,r2,x,y,z,s] to include the spin variable. If the calculation needs
dynamic spins, then a flag will be set called isdynspin.
In some functions, such as translateElectron, it now checks the size of
the translation vector to determine the size, rather than enforcing it
to be 3. Config_point needed changes to the MPI in order to read in and
store the spin variable as well