Molecular system and sample
Finished up the work on the molecular system and molecular
sample. Uses isdynspin variable to determine if it is to be a dynamic
spin calculation. If so, the arrays carrying around the positions
and distances now carry about the spin variable (or potentially
variables for more complicated spin representations). Currently only
using the exp(+is) or exp(-is). Those are evaluated in the orbitals
section of the code, so this passes around the s variable in those
expressions. The spin is initialized between (0,2pi)