Finished Slat_wf_data.*
Finished implementing Slat_wf_data.{h,cpp}. If there is only one type of
spin, then will only hold one type of determinant instead of spin up and
spin down. Additionally, had to make a change in the
system/Molecular_system.h in order to help the wf determine if it is
supposed to be a determinant of only one spin-type. If it sees
sys->nelectrons(1) == -1, then it means that we are only carrying around
one type of spin