Finished wavefunction/Slat_wf.h
The slater_wf file now has an additional flag. If it finds from the
system that there is only one type of spin, namely NSPIN { N } instead
of NSPIN { NUP NDN }, then it knows that we only have 1 determinant
instead of 2, i.e., it does not decompose into a product of spin up and
spin down. It only stores one determinant instead of 2 if we have a
dynamic spin calculation.