Changed such that we use eith NSPIN or DYN_SPIN
If it sees DYN_SPIN in .sys, then we set isdynspin = 1 and passed around
in the code. This allows many of the places in the code that checks
sys->nelectons(0) + sys->nelectrons(1) to still give the total number
electrons, whereas before I had chosen to only store one type of spin if
we have a dynamic spin...i.e., sys->nelectrons(1) wouldn't work because
there was only one type of spin.