System has dynspin check
A boolean variable to check for DYNSPIN in the .sys file
If so, it sets it to true.
If it is true, then when it generates the sample, it generates a
Molecular_sample_dynspin sample, which has the necessary changes. This
is just to get it working. Can change later