Config changes spins
Config file write and read now works for spins. Had to add boolean into
sample_point in order for Config_save_point to know that there is a spin
variable to save

MPI works

NOTE: I did not implement this for th binary files, it looks like binary
file is only used in the Post_process method. Will add if it is ever
needed